Really messy cleanup.  This reverts part of #16640 - reintroducing codeql_config and the usage of it in the github workflow.  DO NOT MERGE THIS INTO MAIN.
Keeping this old branch around as a reminder of hubris and the inexorable weight of icarus while I see if it can be turned into something else.
I'm forced to make a commit before branches are cleaned up, so here we are.

+++ pkg/jmxfetch/jmxfetch_windows.go
@@ -4,7 +4,6 @@
 // Copyright 2016-present Datadog, Inc.
 
 //go:build jmx
-// +build jmx
 
 package jmxfetch
 
@@ -18,9 +17,11 @@ import (
 func (j *JMXFetch) Stop() error {
 	var stopChan chan struct{}
 
-	err := j.cmd.Process.Kill()
-	if err != nil {
-		return err
+	if j.cmd.Process != nil {
+		err := j.cmd.Process.Kill()
+		if err != nil {
+			return err
+		}
 	}
 
 	if j.managed {
@@ -30,7 +31,7 @@ func (j *JMXFetch) Stop() error {
 		stopChan = make(chan struct{})
 
 		go func() {
-			j.Wait()
+			_ = j.Wait()
 			close(stopChan)
 		}()
 	}
